requirements.txt
Django==1.11.6
djangorestframework==3.9.4
gunicorn==19.9.0
pkg-resources==0.0.0
pytz==2019.1
